Fix for address ambiguity problem.
An incoming connection will be recorded with its ephemeral port. As such
if we want to send a peer some data, we finish in making another
connection, because we are unaware of the existence of it.

I solved the problem by making the 'client' send, immediately after the
connection has been created, its IP/port pair. The connections database
gets so far updated with two entries mapped on the same file descriptor.